pyproject.toml
[project]
name = "imdb-movie-recomendation"
version = "0.1.0"
description = "Movie, series, and animation recommendation engine based on IMDB rating history"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pandas>=2.2.0",
    "lightgbm>=4.5.0",
    "scikit-learn>=1.6.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0.0",
    "numpy>=2.0.0",
    "playwright>=1.58.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP"]

[tool.ruff.lint.per-file-ignores]
"app/services/model.py" = ["N806"]

[tool.pytest.ini_options]
pythonpath = ["."]

[project.scripts]
lint = "ruff check app/"
format = "ruff format app/"
